+++ src/reader/types.py
@@ -1,7 +1,6 @@
 import dataclasses
 import re
 import traceback
-import warnings
 from dataclasses import dataclass
 from datetime import datetime
 from types import MappingProxyType
@@ -663,21 +662,6 @@ class UpdatedFeed:
     #: (entries that existed in storage,
     #: but had different data than the corresponding feed file entry.)
     modified: int
-
-    @property
-    def updated(self) -> int:
-        """Deprecated alias for :attr:`UpdatedFeed.modified`.
-
-        .. deprecated: 1.19
-
-        """
-        warnings.warn(
-            "UpdatedFeed.updated is deprecated "
-            "and will be removed in reader 2.0. "
-            "Use UpdatedFeed.modified instead.",
-            DeprecationWarning,
-        )
-        return self.modified
 
 
 class UpdateResult(NamedTuple):